Move sandbox page to dynamic route
