+++ app/projects/[id]/page.tsx
